Refactor package names and update project dependencies.
fix: Renamed packages from "com.example" to "com.ABETAppTeam" across the project to reflect the application branding. Removed unused JSP-related dependencies and added exclusions to resolve conflicts. Cleaned up `.vscode/settings.json` and updated Maven plugin configurations for consistency.
