+++ src/main/java/com/ABETAppTeam/ProfessorServlet.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.ABETAppTeam;
 
 import java.io.IOException;
