+++ src/main/webapp/WEB-INF/web.xml
@@ -11,22 +11,22 @@
     <!-- Servlet Definitions -->
     <servlet>
         <servlet-name>IndexServlet</servlet-name>
-        <servlet-class>com.example.IndexServlet</servlet-class>
+        <servlet-class>com.ABETAppTeam.IndexServlet</servlet-class>
     </servlet>
     
     <servlet>
         <servlet-name>AdminServlet</servlet-name>
-        <servlet-class>com.example.AdminServlet</servlet-class>
+        <servlet-class>com.ABETAppTeam.AdminServlet</servlet-class>
     </servlet>
     
     <servlet>
         <servlet-name>ProfessorServlet</servlet-name>
-        <servlet-class>com.example.ProfessorServlet</servlet-class>
+        <servlet-class>com.ABETAppTeam.ProfessorServlet</servlet-class>
     </servlet>
     
     <servlet>
         <servlet-name>TestServlet</servlet-name>
-        <servlet-class>com.example.TestServlet</servlet-class>
+        <servlet-class>com.ABETAppTeam.TestServlet</servlet-class>
     </servlet>
 
     <!-- Servlet Mappings -->
